+++ src/platform/win32/win32_directory_enumerator.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "application/directory_enumerator.h"
+
+#include <memory>
+#include <string_view>
+
+namespace gitman::win32 {
+    [[nodiscard]] directory_listing enumerate_directory(std::u8string_view absolute_directory) noexcept;
+
+    [[nodiscard]] std::unique_ptr<directory_enumerator> make_directory_enumerator();
+} // namespace gitman::win32